requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bash=5.1.16=hd863ed2_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
curl=7.85.0=h5eee18b_0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=hef1e5e3_1
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
gdbm=1.18=hd4cb3f1_4
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
git=2.34.1=pl5262hc120c5b_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.9.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=10.0.1=default_hb85057a_2
libcurl=7.85.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.6.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.23.3=py38h14f4228_1
numpy-base=1.23.3=py38h31eccc5_1
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pcre2=10.37=he7ceb23_1
perl=5.34.0=h5eee18b_2
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
ply=3.11=py38_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hdb3f193_5
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
setuptools=65.5.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
typing_extensions=4.3.0=py38h06a4308_0
unzip=6.0=h611a1e1_0
urllib3=1.26.12=py38h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
